requirements.txt
fastapi
uvicorn
pydantic
motor
python-jose
passlib
bcrypt==4.0.1
python-decouple
sqlalchemy
psycopg2-binary
python-multipart

# required for stable diffusion
diffusers==0.25.1
torch==2.0.0
torchvision==0.15.1
transformers==4.37.1
accelerate
safetensors
einops
onnxruntime-gpu
spaces==0.19.4
omegaconf
peft
huggingface-hub==0.20.2
opencv-python
insightface
gradio
controlnet_aux
gdown
mediapipe
xformers